# docs/postman_OpenApiV3_response.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Car-Rest-Service API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/manufacturers/{manufacturerName}": {
      "get": {
        "tags": [
          "manufacturers"
        ],
        "summary": "Get manufacturer by name",
        "description": "Retrieve a specific manufacturer by its unique name",
        "operationId": "getManufacturerByName",
        "parameters": [
          {
            "name": "manufacturerName",
            "in": "path",
            "description": "Name of the manufacturer",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Toyota"
          }
        ],
        "responses": {
          "200": {
            "description": "Manufacturer retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MakeDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid manufacturerName format"
          },
          "404": {
            "description": "Manufacturer not found with provided name"
          }
        }
      },
      "put": {
        "tags": [
          "manufacturers"
        ],
        "summary": "Update manufacturer by name",
        "description": "Update an existing manufacturer. Requires authentification.",
        "operationId": "updateManufacturer",
        "parameters": [
          {
            "name": "manufacturerName",
            "in": "path",
            "description": "Name of the manufacturer",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Toyota"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MakeDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Manufacturer updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MakeDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data or manufacturerName format"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token"
          },
          "404": {
            "description": "Make not found with provided name"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "manufacturers"
        ],
        "summary": "Delete manufacturer by name",
        "description": "Delete a specific manufacturer. Requires authentification.",
        "operationId": "deleteManufacturer",
        "parameters": [
          {
            "name": "manufacturerName",
            "in": "path",
            "description": "Name of the manufacturer",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Toyota"
          }
        ],
        "responses": {
          "204": {
            "description": "Manufacturer deleted successfully"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token"
          },
          "404": {
            "description": "Manufacturer not found with provided name"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/manufacturers/{manufacturerName}/models/{modelName}": {
      "get": {
        "tags": [
          "manufacturers"
        ],
        "summary": "Get model by manufacturer and model name",
        "description": "Retrieve a specific model for a manufacturer by model name",
        "operationId": "getModelByManufacturerAndName",
        "parameters": [
          {
            "name": "manufacturerName",
            "in": "path",
            "description": "Name of the manufacturer",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Toyota"
          },
          {
            "name": "modelName",
            "in": "path",
            "description": "Name of the model",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Camry"
          }
        ],
        "responses": {
          "200": {
            "description": "Model retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Manufacturer or model not found with provided name"
          }
        }
      },
      "put": {
        "tags": [
          "manufacturers"
        ],
        "summary": "Update model by manufacturer and model name",
        "description": "Update an existing model by a manufacturer. Requires authentification.",
        "operationId": "updateModelByManufacturerAndName",
        "parameters": [
          {
            "name": "manufacturerName",
            "in": "path",
            "description": "Name of the manufacturer",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Toyota"
          },
          {
            "name": "modelName",
            "in": "path",
            "description": "Name of the model",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Camry"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Model updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data or parameter format"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token"
          },
          "404": {
            "description": "Manufacturer or model not found with provided name"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "manufacturers"
        ],
        "summary": "Delete model by manufacturer and model name",
        "description": "Delete a specific model by a manufacturer. Requires authentification.",
        "operationId": "deleteModelByManufacturerAndName",
        "parameters": [
          {
            "name": "manufacturerName",
            "in": "path",
            "description": "Name of the manufacturer",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Toyota"
          },
          {
            "name": "modelName",
            "in": "path",
            "description": "Name of the model",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Camry"
          }
        ],
        "responses": {
          "204": {
            "description": "Model deleted successfully"
          },
          "400": {
            "description": "Invalid parameter format"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token"
          },
          "404": {
            "description": "Manufacturer or model not found with provided name"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/manufacturers": {
      "get": {
        "tags": [
          "manufacturers"
        ],
        "summary": "Get all manufacturers with pagination",
        "description": "Retrieve a paginated list of all manufacturers",
        "operationId": "getAllManufacturers",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "description": "Pagination information",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Manufacturers retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination parameters"
          }
        }
      },
      "post": {
        "tags": [
          "manufacturers"
        ],
        "summary": "Create a new manufacturer",
        "description": "Create a new manufacturer record. Requires authentification.",
        "operationId": "createManufacturer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MakeDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Manufacturer created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MakeDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token"
          },
          "409": {
            "description": "Manufacturer already exists"
          }
        }
      }
    },
...
    "/api/v1/manufacturers/id/{makeId}": {
      "get": {
        "tags": [
          "manufacturers"
        ],
        "summary": "Get manufacturer by id",
        "description": "Retrieve a specific manufacturer by its unique id",
        "operationId": "getManufacturerById",
        "parameters": [
          {
            "name": "makeId",
            "in": "path",
            "description": "Unique identifier of the manufacturer",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Manufacturer retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MakeDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid makeId format"
          },
          "404": {
            "description": "Manufacturer not found with provided id"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "MakeDto": {
        "type": "object",
        "properties": {
          "makeId": {
            "type": "integer",
            "format": "int32"
          },
          "make": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [
          "make"
        ]
      },
      "ModelDto": {
        "type": "object",
        "properties": {
          "modelId": {
            "type": "integer",
            "format": "int32"
          },
          "make": {
            "$ref": "#/components/schemas/MakeDto"
          },
          "model": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [
          "make",
          "model"
        ]
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Page": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {}
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}